# [feature] Files Page Optimizations [OSF-8677]