--- a/website/static/js/fileViewTreebeard.js
+++ b/website/static/js/fileViewTreebeard.js
@@ -61,14 +61,6 @@ function FileViewTreebeard(data) {
                 width: '100%'
             }];
         },
-        ontogglefolder : function (tree) {
-            Fangorn.DefaultOptions.ontogglefolder.call(this, tree);
-            var containerHeight = this.select('#tb-tbody').height();
-            if (!this.options.naturalScrollLimit){
-                this.options.showTotal = Math.floor(containerHeight / this.options.rowHeight) + 1;
-            }
-            this.redraw();
-        },
         lazyLoadOnLoad: function(tree, event) {
             var tb = this;
             Fangorn.DefaultOptions.lazyLoadOnLoad.call(tb, tree, event);