--- a/website/util/rubeus.py
+++ b/website/util/rubeus.py
@@ -11,6 +11,8 @@ from framework import sentry
 from framework.auth.decorators import Auth
 
 from django.apps import apps
+from django.db import connection
+from django.db.models import Exists, OuterRef
 
 from website import settings
 from website.util import paths
@@ -163,32 +165,13 @@ class NodeFileCollector(object):
         """Return the Rubeus.JS representation of the node's file data, including
         addons and components
         """
-        root = self._serialize_node(self.node)
+        print(len(connection.queries))
+        root = self._get_nodes(self.node)
         return [root]
 
-    def _collect_components(self, node, visited):
-        rv = []
-        if not node.can_view(self.auth):
-            return rv
-        for child in node.get_nodes(is_deleted=False):
-            if not child.can_view(self.auth):
-                if child.primary:
-                    for desc in child.find_readable_descendants(self.auth):
-                        visited.append(desc.resolve()._id)
-                        rv.append(self._serialize_node(desc, visited=visited, parent=node))
-            elif child.resolve()._id not in visited:
-                visited.append(child.resolve()._id)
-                rv.append(self._serialize_node(child, visited=visited, parent=node))
-        return rv
-
-    def _get_node_name(self, node):
+    def _get_node_name(self, node, can_view):
         """Input node object, return the project name to be display.
         """
-        NodeRelation = apps.get_model('osf.NodeRelation')
-        is_node_relation = isinstance(node, NodeRelation)
-        node = node.child if is_node_relation else node
-        can_view = node.can_view(auth=self.auth)
-
         if can_view:
             node_name = sanitize.unescape_entities(node.title)
         elif node.is_registration:
@@ -202,26 +185,18 @@ class NodeFileCollector(object):
 
         return node_name
 
-    def _serialize_node(self, node, visited=None, parent=None):
-        """Returns the rubeus representation of a node folder.
-        """
-        visited = visited or []
-        visited.append(node.resolve()._id)
+    def _serialize_node(self, node, parent=None, children=[]):
+        is_pointer = parent and node.linked_node
         can_view = node.can_view(auth=self.auth)
-        if can_view:
-            children = self._collect_addons(node) + self._collect_components(node, visited)
-        else:
-            children = []
-
-        is_pointer = parent and parent.has_node_link_to(node)
+        can_edit = node.has_write_perm if hasattr(node, 'has_write_perm') else node.can_edit(auth=self.auth)
 
         return {
             # TODO: Remove safe_unescape_html when mako html safe comes in
-            'name': self._get_node_name(node),
+            'name': self._get_node_name(node, can_view),
             'category': node.category,
             'kind': FOLDER,
             'permissions': {
-                'edit': node.can_edit(self.auth) and not node.is_registration,
+                'edit': can_edit and not node.is_registration,
                 'view': can_view,
             },
             'urls': {
@@ -231,9 +206,28 @@ class NodeFileCollector(object):
             'children': children,
             'isPointer': is_pointer,
             'isSmartFolder': False,
-            'nodeType': node.project_or_component,
-            'nodeID': node.resolve()._id,
+            'nodeType': 'component' if parent else 'project',
+            'nodeID': node._id,
         }
+
+    def _get_nodes(self, node):
+        AbstractNode = apps.get_model('osf.AbstractNode')
+        Contributor = apps.get_model('osf.Contributor')
+        NodeRelation = apps.get_model('osf.NodeRelation')
+
+        data = []
+        if node.can_view(auth=self.auth):
+            serialized_addons = self._collect_addons(node)
+            linked_node_sqs = NodeRelation.objects.filter(parent=node, is_node_link=True)
+            has_write_perm_sqs = Contributor.objects.filter(node=OuterRef('pk'), write=True, user=self.auth.user)
+            children = (AbstractNode.objects
+                        .filter(is_deleted=False, _parents__parent=node)
+                        .annotate(linked_node=Exists(linked_node_sqs))
+                        .annotate(has_write_perm=Exists(has_write_perm_sqs))
+                        )
+            serialized_children = [self._serialize_node(child, parent=node) for child in children]
+            data = serialized_addons + serialized_children
+        return self._serialize_node(node, children=data)
 
     def _collect_addons(self, node):
         rv = []